+++ src/product-sku/index.ts
@@ -1,2 +1,0 @@
-export { ProductSkuManager } from './productSku'
-export * from './types'